Solucion a problema con imagenes

+++ src/main/java/com/edwin/galeriademo/controller/fotoController.java
@@ -7,10 +7,12 @@ import com.edwin.galeriademo.service.fotoService;
 import org.apache.tomcat.util.http.fileupload.FileUpload;
 import org.slf4j.*;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.io.IOException;
 import java.util.Optional;
@@ -30,16 +32,19 @@ public class fotoController {
     @GetMapping("")
     public String show(Model model) {
         model.addAttribute("fotos", fotoService.findAll());
+
         return "fotos/show";
     }
 
     @GetMapping("/create")
     public String create() {
-    return "fotos/create";}
+        return "fotos/create";
+    }
 
     @PostMapping("/save")
     public String save(foto foto, @RequestParam("img") MultipartFile file) throws IOException {
         LOGGER.info("Saving foto: {}", foto);
+
         usuario u= new usuario(1,"","","");
         foto.setUsuario(u);
 
@@ -47,12 +52,13 @@ public class fotoController {
         if (foto.getId() ==null){
             String nombrefoto = upload.saveImage(file);
             foto.setImagen(nombrefoto);
+
         } else{
 
         }
 
         fotoService.save(foto);
-        return "redirect:/fotos/";
+        return "redirect:/fotos";
     }
     @GetMapping("/edit/{id}")
     public String edit(@PathVariable Integer id, Model model) {
@@ -66,18 +72,50 @@ public class fotoController {
     }
 
 
-
-
     @PostMapping("/update")
-    public String update(foto foto) {
+    public String update(foto foto, @RequestParam("img") MultipartFile file) throws IOException {
         fotoService.update(foto);
-        return "redirect:/fotos/";
+
+        if (file.isEmpty()) {
+            foto p = new foto();
+            p=fotoService.get(foto.getId()).get();
+            foto.setImagen(p.getImagen());
+        } else { // cuando se edita la imagen
+
+            foto p = new foto();
+            p=fotoService.get(foto.getId()).get();
+            // eliminar cuando no sea la imagen por defecto
+
+            if (p.getImagen().equals("default.jpg")) {
+                upload.deleteImage(p.getImagen());
+            }
+
+            String nombrefoto = upload.saveImage(file);
+            foto.setImagen(nombrefoto);
+        }
+
+        return "redirect:/fotos";
     }
 
     @GetMapping("/delete/{id}")
     public String delete(@PathVariable Integer id) {
-        fotoService.delete(id);
-        return "redirect:/fotos/";
+        Optional<foto> optionalFoto = fotoService.get(id);
+
+        if (optionalFoto.isPresent()) {
+            foto p = optionalFoto.get();
+
+            // eliminar cuando no sea la imagen por defecto
+            if (!p.getImagen().equals("default.jpg")) {
+                upload.deleteImage(p.getImagen());
+            }
+
+            fotoService.delete(id);
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Foto no encontrada");
+        }
+
+        return "redirect:/fotos";
     }
+
 
 }
